Split the BuildStage class
